utf8.c: swash_to_invlist() handle EXTRAS
This function has not been able to handle what are called EXTRAS in
its input.  These are things like:

    !utf8::InHiragana
    -utf8::InKatakana
    +utf8::IsCn

besides the normal list of ranges.

This commit allows this function to handle all the same constructs as
the regular swash input function, from which most of the new code was
copied.